+++ game/scrabble.py
@@ -22,6 +22,9 @@ class ScrabbleGame:
             self.turn = 0
         
     def validate_and_put_word (self,word, location, orientation):
+        # palabra = []
+        # for i in word:
+        #     palabra.append(i)
         playerTiles = self.players[self.turn].tiles
         v1 = self.board.validate_word_inside_board(word,location, orientation)
         if not v1:
@@ -60,11 +63,11 @@ class ScrabbleGame:
                 M += 1
             elif orientation == 'V':
                 N += 1
-
             
         self.board.put_words(wordToColocate, location, orientation)
         playerTiles.extend(self.bag_tiles.take(len(word)))  # User takes tiles from the bag
-        # self.players[self.turn].score += Tools().calculate_word_value(wordToColocate)
+        word_cells = self.board.cells_of_word_in_board(word, location , orientation)
+        self.players[self.turn].score += Tools().calculate_word_value(word_cells)
         return ('Word succesfully colocated.')